[beta] may revert: consolidating duplicate logic into hook and main message file
